requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via -r requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via uvicorn
fastapi==0.115.6
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pydantic==2.10.5
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pytest==8.3.4
    # via pytest-asyncio
pytest-asyncio==0.25.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via -r requirements.in
